backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.7.0
slowapi>=0.1.9
httpx>=0.27.0
PyJWT>=2.8.0
python-multipart>=0.0.9
Pillow>=10.3.0
pytesseract>=0.3.13
